gzh_task6/requirements.txt
openai>=1.50.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
rich>=13.0.0